Implement alternate_keys
This implements an alternate_keys config argument which lets you specify
a set of keys to check. This makes it possible to be backwards compatible
with configuration key changes.

Additionally, this adds handling for a special "root:" key prefix
which indicates to look at the root of the configuration rather than
in the current namespace. This makes it possible to allow for shared
configuration between components. For example, hosts, usernames and
passwords for services.

Fixes #19