Fix listify when passed None
Prior to this fix, if you called listify(None), it'd return something
unhelpful. Now it returns the empty list.

The thinking here is that if listify gets None as an argument, it's
as if it got a string as an argument so we turn it into a list. But
None indicates no value--not a list of one value--so we turn it into
an empty list.