Fix nested BoundConfig case
In a situation where there are nested BoundConfigs in the config chain, then
parents configs will raise a KeyError because that key isn't in *their*
options.

This fixes that by removing BoundConfigs from the chain.

This also adds __repr__ which will make things easier to debug in complex
config scenarios.