Initial scaffolding for type checking (#155)
This adds the initial scaffolding for type checking and fixes existing
issues mypy pointed out.

docutils and configobj don't have type annotations available, so they're
ignored for now. We can change that later when they're available.

This also adds a test for qualname and removes some of the qualname code
that seems like it's from the Python 2 days.