Reworked ConfigurationError msg and switched to f-strings
I decided instead of bending over backwards for a clear, but general,
way of referring to a namespace/key combination in the error output, I'd
go with the structure that Everett uses for keys in the process
environment and env files. I claim this is the most common usage for
Everett, so we should cater the messages that way and further reduce
confusion users have between the general formatting and what might be in
the project documentation.

To make that work with other systems like INI files where namespaces are
implemented as INI sections, I added a msg_builder argument to
ConfigManager. I think that'll be good enough.

While doing that, I switched a lot of composed strings to f-strings and
took out some Python <=3.5 code that was still around.