Overhaul sphinx extension; support module configuration docs (#172)
This overhauls the Sphinx extension.

This now supports manually documenting configuration using
``everett:component`` and ``everett:option`` directives.

This adds ``:everett:component:`` and ``:everett:option:`` roles for
linking to specific configuration in the docs.

It also addsh ``autocomponentconfig`` and ``automoduleconfig``
directives for automatically generating documentation.

When using these directives, items are added to the index and everything
is linkable making it easier to find and talk to users about specific
configuration items.