Overhaul documentation
This simplifies and cleans up the README and also moves the code into
separate files that are injected (with output) into the README using
cog.

This breaks up the monolithic configuration chapter into a bunch of
smaller topical chapters.

This moves all the example code into examples/ and reworks the
configuration chapter to use cog to insert output from running code.

This fixes some of the examples so they're less wonky and more
self-contained. It also adds linting for them.

Overall this should improve the documentation for people looking at
Everett for the first time and for people looking to do specific things.
It also reduces the likelihood that the example code is wrong or
out-of-date. It also reduces the overall maintenance burden.