Add raise_configuration_error convenience method (#185)
This adds a `raise_configuration_error` convenience method to
`ConfigManager` making it easier for developers to raise a configuration
error that will pick up the configuration documentation in the message.
This is useful for cases where the application needs to validate values
of configuration options that are connected.

For example, if you want to require the user to either set the host and
port or neither, you can validate it and use `raise_configuration_error`
if the user does something you didn't want.