Fix sphinxext related issues
Sphinx 6.2.1 is the last version of Sphinx that supports Python 3.8, so
this updates the dev requirements to use that.

There were changes in the ast module, so this updates the sphinxext code
to stop using deprecated things.

This adds "sphinx" as an extra and adds that to the relevant
documentation.